Put alarms in a file
