+++ lib/cfn.js
@@ -4,6 +4,7 @@ const webhooks = require('./events/webhook');
 const snsEvent = require('./events/sns');
 const cloudWatchEvent = require('./events/cloudwatch');
 const roles = require('./roles');
+const alarms = require('./alarms');
 
 module.exports = buildFunctionTemplate; // returns full template
 module.exports.embed = buildFunction; // returns only lambda resources for template merge
@@ -176,7 +177,7 @@ function buildFunction(options) {
     throw new Error('Function name is required');
 
   var parameters = buildParameters(options);
-  var defaultAlarms = buildServiceAlarms(options);
+  var defaultAlarms = alarms.buildServiceAlarms(options);
   var lambda = buildLambda(options);
   var role = roles.buildRole(options);
   var cloudwatchEvent = {};
@@ -389,75 +390,6 @@ function buildLambda(options) {
   return fn;
 }
 
-
-module.exports.buildServiceAlarms = buildServiceAlarms;
-function buildServiceAlarms(options) {
-  var alarms = {
-    Parameters: {},
-    Resources: {},
-    Variables: {}
-  };
-
-  var defaultAlarms = [
-    {
-      AlarmName: 'Errors',
-      MetricName: 'Errors',
-      ComparisonOperator: 'GreaterThanThreshold'
-    },
-    {
-      AlarmName: 'NoInvocations',
-      MetricName: 'Invocations',
-      ComparisonOperator: 'LessThanThreshold'
-    }
-  ];
-
-  defaultAlarms.forEach(function(alarm) {
-    alarms.Resources[options.name + 'Alarm' + alarm.AlarmName] = {
-      Type: 'AWS::CloudWatch::Alarm',
-      Properties: {
-        EvaluationPeriods: '5',
-        Statistic: 'Sum',
-        Threshold: '0',
-        AlarmDescription: 'https://github.com/mapbox/lambda-cfn/blob/master/alarms.md#' + alarm.AlarmName,
-        Period: '60',
-        AlarmActions: [cf.ref('ServiceAlarmSNSTopic')],
-        Namespace: 'AWS/Lambda',
-        Dimensions: [
-          {
-            Name: 'FunctionName',
-            Value: cf.ref(options.name)
-          }
-        ],
-        ComparisonOperator: alarm.ComparisonOperator,
-        MetricName: alarm.MetricName
-      }
-    };
-  });
-
-  alarms.Parameters = {
-    ServiceAlarmEmail:{
-      Type: 'String',
-      Description: 'Service alarm notifications will send to this email address'
-    }
-  };
-
-  alarms.Resources.ServiceAlarmSNSTopic = {
-    Type: 'AWS::SNS::Topic',
-    Properties: {
-      TopicName: cf.join('-', [cf.stackName, 'ServiceAlarm']),
-      Subscription: [
-        {
-          Endpoint: cf.ref('ServiceAlarmEmail'),
-          Protocol: 'email'
-        }
-      ]
-    }
-  };
-
-  alarms.Variables.ServiceAlarmSNSTopic = cf.ref('ServiceAlarmSNSTopic');
-  return alarms;
-}
-
 module.exports.buildSnsDestination = buildSnsDestination;
 function buildSnsDestination(options) {
 
@@ -513,3 +445,4 @@ module.exports.buildWebhookEvent = webhooks.buildWebhookEvent;
 module.exports.buildSnsEvent = snsEvent.buildSnsEvent;
 module.exports.buildCloudwatchEvent = cloudWatchEvent.buildCloudwatchEvent;
 module.exports.buildRole = roles.buildRole;
+module.exports.buildServiceAlarms = alarms.buildServiceAlarms;
